Add settings control
