+++ apps/studio/src/routes/editor/TopBar/Publish/index.tsx
@@ -1,14 +1,17 @@
 import { DropdownMenu, DropdownMenuContent } from '@onlook/ui/dropdown-menu';
 import { observer } from 'mobx-react-lite';
+import { useState } from 'react';
 import { PublishDropdown } from './Dropdown';
 import { PublishButton } from './TriggerButton';
 
 const Publish = observer(() => {
+    const [isOpen, setIsOpen] = useState(false);
+
     return (
-        <DropdownMenu>
+        <DropdownMenu open={isOpen} onOpenChange={setIsOpen}>
             <PublishButton />
             <DropdownMenuContent align="end" className="w-96 p-0">
-                <PublishDropdown />
+                <PublishDropdown setIsOpen={setIsOpen} />
             </DropdownMenuContent>
         </DropdownMenu>
     );
